requirements.txt
biopython==1.81
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
filelock==3.12.2
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.2
platformdirs==4.1.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
tokenizers==0.13.3
torch==1.13.1
tqdm==4.66.1
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.6
virtualenv==20.25.0
zipp==3.15.0
